Добавлено автообнаружение hx710b и vl53l0x в гнезде US, а так-же тестовое определение наличия фоторезистора
