+++ platformio.ini
@@ -29,7 +29,7 @@ lib_deps =
 
 ; Строку ниже нужно разкоментировать для прошивки ESP по Wi-Fi
 ; 192.168.XX.XX заменить на свой IP адрес, который ESP получит после удачной прошивки через консольный кабель
-upload_port = wegabox.local
+upload_port = 192.168.237.51
 upload_flags = "--timeout=10"
 upload_protocol = espota
