+++ src/dev/hx710b/setup.h
@@ -1,15 +1,18 @@
 #if c_HX710B == 1
-while (xSemaphoreTake(xSemaphoreX, (TickType_t)1) == pdFALSE);
+while (xSemaphoreTake(xSemaphoreX, (TickType_t)1) == pdFALSE)
+  ;
 
-  if ( !HX710B_Press.init() )
-  { 
-    syslog_ng("HX710B not Found !");    
-  }
-  // else
-  // {
-        xSemaphoreGive(xSemaphoreX);    
+uint32_t HX710B_probe = -1;
+if (HX710B_Press.init())
+  HX710B_Press.read(&HX710B_probe, 1000UL);
 
-xTaskCreate(TaskHX710B,"TaskHX710B",15000,NULL,0,NULL);
-syslog_ng("HX710B add Task");
-//  }
-#endif //c_HX710B
+if (HX710B_probe != -1)
+{
+  xTaskCreate(TaskHX710B, "TaskHX710B", 15000, NULL, 0, NULL);
+  syslog_ng("HX710B: add Task");
+}
+else
+  syslog_err("HX710B: The sensor is not detected");
+xSemaphoreGive(xSemaphoreX);
+
+#endif // c_HX710B
