+++ src/dev/pr/tasks.h
@@ -2,8 +2,7 @@
 #if c_PR == 1
 void TaskPR(void *parameters)
 {
-  adc1_config_width(ADC_WIDTH_BIT_12);
-  adc1_config_channel_atten(PR_AnalogPort, ADC_ATTEN_DB_11);
+
   for (;;)
   {
     if (OtaStart == true)
